+++ source/extensions/omni.isaac.lab_assets/omni/isaac/lab_assets/__init__.py
@@ -26,6 +26,7 @@ __version__ = ISAACLAB_ASSETS_METADATA["package"]["version"]
 # Configuration for different assets.
 ##
 
+# robots
 from .allegro import *
 from .ant import *
 from .anymal import *
@@ -37,5 +38,8 @@ from .quadcopter import *
 from .ridgeback_franka import *
 from .sawyer import *
 from .shadow_hand import *
+
+# sensors
+from .stereolabs import *
 from .unitree import *
 from .universal_robots import *